+++ alfresco-helloworld-transformer-amp/helloworld-amp/src/main/amp/config/alfresco/module/helloworld-amp/module-context.xml
@@ -1,0 +1,30 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<!--
+	Licensed to the Apache Software Foundation (ASF) under one or more
+	contributor license agreements. 
+	The ASF licenses this file to You under the Apache License, Version 2.0
+	(the "License"); you may not use this file except in compliance with
+	the License.  You may obtain a copy of the License at
+	
+	http://www.apache.org/licenses/LICENSE-2.0
+	
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+	
+-->
+
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans
+          http://www.springframework.org/schema/beans/spring-beans-3.0.xsd">
+
+	<!-- This is filtered by Maven at build time, so that module name is single sourced. -->
+	<!-- Note. The bootstrap-context.xml file has to be loaded first.
+				Otherwise your custom models are not yet loaded when your service beans are instantiated and you
+				cannot for example register policies on them. -->
+	<import resource="classpath:alfresco/module/${project.artifactId}/context/hello-world-context.xml" />
+
+</beans>
