+++ alfresco-helloworld-transformer-amp/packaging/docker-compose.yml
@@ -1,0 +1,43 @@
+version: "3"
+
+services:
+    alfresco:
+        image: alfresco/acs-hello:latest
+        environment:
+            CATALINA_OPTS : "
+                -agentlib:jdwp=transport=dt_socket,address=*:8000,server=y,suspend=n
+                "
+            JAVA_OPTS : "
+                -Dalfresco.port=8082
+                -Daos.baseUrlOverwrite=http://localhost:8082/alfresco/aos
+                -Ddb.driver=org.postgresql.Driver
+                -Ddb.username=alfresco
+                -Ddb.password=alfresco
+                -Ddb.url=jdbc:postgresql://postgres:5432/alfresco
+                -Dindex.subsystem.name=solr6
+                -Dalfresco.restApi.basicAuthScheme=true
+                -Dmessaging.broker.url=\"failover:(tcp://activemq:61616)?timeout=3000&jms.useCompression=true\"
+                -Dtransform.service.enabled=false
+                -Dlocal.transform.service.enabled=false
+                -Dlegacy.transform.service.enabled=true
+                "
+        ports:
+            - 8082:8080
+            - 8000:8000
+
+    postgres:
+        image: library/postgres:11.4
+        environment:
+            - POSTGRES_PASSWORD=alfresco
+            - POSTGRES_USER=alfresco
+            - POSTGRES_DB=alfresco
+        ports:
+            - 5432:5432
+
+    activemq:
+        image: alfresco/alfresco-activemq:5.15.8
+        ports:
+            - 8161:8161 # Web Console
+            - 5672:5672 # AMQP
+            - 61616:61616 # OpenWire
+            - 61613:61613 # STOMP
