+++ alfresco-helloworld-transformer-engine/src/main/resources/application-default.yaml
